Added an unexported method for uploadLargeFile in fileUploader and added a decision statement in the exported method to decide when to use it

+++ goblaze/accountauthorization/accountauthorization.go
@@ -13,8 +13,8 @@ type AuthorizeAccountResponse struct {
 	Allowed                 TokenCapabilities
 	APIURL                  string
 	DownloadURL             string
-	RecommendedPartSize     int
-	AbsoluteMinimumPartSize int
+	RecommendedPartSize     int64
+	AbsoluteMinimumPartSize int64
 }
 
 type TokenCapabilities struct {
